pyscf: a correlated method keeps its HF reference family; per-root flags travel as integers
Two defects the first real runs of the v5 driver found, repaired where they
were made. The result validator derived the expected reference family only
for ``ab_initio: hf`` and returned none for ``mp2``/``ccsd``/``ccsd(t)``, so
every correlated artifact was refused with ``reference_family_mismatch``
against the RHF mean field it correctly ran on; every ab initio method
runs the same mean field and the family of that reference is the fact to
check. And the driver wrote ``excited_state_converged`` as a boolean HDF5
array, which the contract's numeric-dtype rule refuses as it should; the
per-root flags travel as 0/1 integers and the reader restores the booleans.

+++ chemsmart/jobs/pyscf/writer.py
@@ -1032,7 +1032,9 @@ def _run_td(config, mf, results, status, runtime):
     converged = _root_convergence(td)
     manifold = str(config["state_manifold"]).strip().lower()
     results["excitation_energies"] = excitations
-    results["excited_state_converged"] = converged
+    # Physical arrays are integer or floating in this contract; the
+    # per-root flags travel as 0/1 and the reader restores the booleans.
+    results["excited_state_converged"] = converged.astype(int)
     if manifold in ("singlet", "triplet"):
         results["excited_state_multiplicities"] = np.full(
             obtained, 1 if manifold == "singlet" else 3, dtype=int
